Split billing components. Extract common components.
