+++ src/components/TableRowEmptyState/index.ts
@@ -1,0 +1,1 @@
+export { default, Props as TableRowEmptyStateProps } from './TableRowEmptyState';
