+++ src/components/TableRowError/index.ts
@@ -1,0 +1,1 @@
+export { default, Props as TableRowErrorProps } from './TableRowError';
