+++ src/components/TableRowLoading/TableRowLoading.tsx
@@ -1,0 +1,31 @@
+import { StyleRulesCallback, Theme, withStyles, WithStyles } from '@material-ui/core/styles';
+import TableCell from '@material-ui/core/TableCell';
+import TableRow from '@material-ui/core/TableRow';
+import * as React from 'react';
+import CircleProgress from 'src/components/CircleProgress';
+
+type ClassNames = 'root';
+
+const styles: StyleRulesCallback<ClassNames> = (theme: Theme) => ({
+  root: {},
+});
+
+export interface Props {
+  colSpan: number;
+}
+
+type CombinedProps = Props & WithStyles<ClassNames>;
+
+const TableRowLoading: React.StatelessComponent<CombinedProps> = (props) => {
+  return (
+    <TableRow>
+      <TableCell colSpan={props.colSpan}>
+        <CircleProgress noTopMargin />
+      </TableCell>
+    </TableRow>
+  );
+};
+
+const styled = withStyles(styles, { withTheme: true });
+
+export default styled<Props>(TableRowLoading);
