+++ src/components/TableRowLoading/index.ts
@@ -1,0 +1,1 @@
+export { default, Props as TableRowProps } from './TableRowLoading';
