+++ src/features/Account/AccountDetail.tsx
@@ -9,7 +9,8 @@ import { Requestable } from 'src/requestableContext';
 import { getAccountInfo } from 'src/services/account';
 
 import MakeAPaymentPanel from './AccountPanels/MakeAPaymentPanel';
-import RecentBillingActivityPanel from './AccountPanels/RecentBillingActivityPanel';
+import RecentInvoicesPanel from './AccountPanels/RecentInvoicesPanel';
+import RecentPaymentsPanel from './AccountPanels/RecentPaymentsPanel';
 import SummaryPanel from './AccountPanels/SummaryPanel';
 import UpdateContactInformationPanel from './AccountPanels/UpdateContactInformationPanel';
 import UpdateCreditCardPanel from './AccountPanels/UpdateCreditCardPanel';
@@ -146,7 +147,8 @@ export class AccountDetail extends React.Component<CombinedProps, State> {
           <Typography variant="title" className={classes.heading}>Billing Information</Typography>
           <UpdateCreditCardPanel />
           <MakeAPaymentPanel />
-          <RecentBillingActivityPanel />
+          <RecentInvoicesPanel />
+          <RecentPaymentsPanel />
         </AccountProvider >
       </React.Fragment>
     );
