--- src/features/Account/AccountPanels/RecentBillingActivityPanel/index.ts
@@ -1,1 +1,0 @@
-export { default } from './RecentBillingActivityPanel';
