+++ src/features/Account/AccountPanels/RecentInvoicesPanel/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RecentInvoicesPanel';
