+++ src/features/Account/AccountPanels/RecentPaymentsPanel/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RecentPaymentsPanel';
